# server/swagger/swaggerCompany.yaml
...
openapi: 3.0.3
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API
    url: http://localhost:5000/api
info:
  version: "1.0.0"
  title: Company
  description: The API for the task project from Techorely
paths:
...
  '/user':
    get:
      tags:
        - Fetch all users
      description: returns all users
      operationId: getAllUsers
      parameters:
        - in: query
          name: page
          description: number of records to skip
          schema:
            type: integer
            format: int32
        - in: query
          name: limit
          description: max number of records to return
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Fetch all the companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - Create user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Succesful created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    x-swagger-router-controller: userRouter
...
  '/user/{userId}':
    get:
      tags:
        - get user
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: integer
      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-swagger-router-controller: userRouter

    patch:
      tags:
        - update user
      operationId: updateByUserId
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-swagger-router-controller: userRouter

    delete:
      tags:
        - delete user
      operationId: deleteByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not found'
        '404':
          description: User by id is not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error of server'
      x-swagger-router-controller: userRouter
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
        role:
          $ref: '#/components/schemas/Role'
        user_info:
          $ref: '#/components/schemas/UserInfo'
      required:
        - id
        - email
        - password
        - role
        - user_info
        -
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        img:
          type: string
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string